+++ lib/src/main/kotlin/cz/eman/swagger/codegen/SwaggerCodeGenPlugin.kt
@@ -6,6 +6,7 @@ import org.openapitools.codegen.config.CodegenConfigurator
 
 /**
  * @author eMan s.r.o. (vaclav.souhrada@eman.cz)
+ * @since 1.0.0
  */
 open class SwaggerCodeGenPlugin : Plugin<Project> {
     companion object TaskNames {