+++ lib/src/main/kotlin/cz/eman/swagger/codegen/SwaggerCodeGenTask.kt
@@ -4,6 +4,7 @@ import org.openapitools.codegen.config.CodegenConfigurator
 
 /**
  * @author eMan s.r.o. (vaclav.souhrada@eman.cz)
+ * @since 1.0.0
  */
 open class SwaggerCodeGenTask : AbstractSwaggerGenTask() {