+++ CHANGELOG.md
@@ -3,9 +3,12 @@ Change Log
 
 ## 1.1.0 (TBD)
 - Migrated from Swagger to OpenApi generator (https://github.com/OpenAPITools/openapi-generator).
-- Merged Retrofit and Android generator together.
-- Empty data class is defined as a String alias.
+- Merged Retrofit and Room generator together.
+- Empty data class is defined as a String type alias.
 - Changed anonymous enum names from _* to NUMBER_*.
+- Added support for Room 2 (androidx).
+- Added option to change array of composed to array of object (kotlin.Any).
+- Added option to generate primitive aliases.
 
 ## 1.0.0 (2019-06-25)
 - Initial version